fix: ensure .tasks directory exists before writing to it
- Add ensure_directory_exists() utility function to create parent directories
- Fix observation.py to ensure .tasks directory exists before writing backup file
- Centralize DEAD_LETTER_DIRECTORY constant in utils.py
- Remove duplicate queue_dead_letter functions across multiple files
- Add clean wrapper functions queue_failed_request() and retry_failed_requests()
- Update all files to use centralized dead letter management
